migrate detekt to project level
