+++ .gitignore
@@ -13,4 +13,3 @@
 .externalNativeBuild
 .cxx
 local.properties
-/detekt-setting/baseline.xml
