+++ app/build.gradle.kts
@@ -1,5 +1,3 @@
-import io.gitlab.arturbosch.detekt.Detekt
-
 plugins {
     alias(libs.plugins.android.application)
     alias(libs.plugins.kotlin.android)
@@ -42,100 +40,6 @@ android {
     }
 }
 
-// default lint config - only warnings
-android {
-    lint {
-        checkDependencies = true  // Enable Lint for dependencies
-        htmlOutput = file("${layout.buildDirectory.get()}/reports/lint/lint-report.html")
-    }
-}
-
-detekt {
-    toolVersion = libs.versions.detekt.get()
-    config.setFrom(file("$projectDir/detekt-setting/detekt.yml")) // Empty = use defaults detekt rules
-    baseline = file("$projectDir/detekt-setting/baseline.xml")
-}
-
-tasks.withType<Detekt>().configureEach {
-    jvmTarget = JavaVersion.VERSION_21.toString()
-
-    // Disable default failure behavior (critical step!)
-    ignoreFailures = false  // Keep false to ensure build fails
-    autoCorrect = false     // Optional: Disable auto-fixing
-    buildUponDefaultConfig = true // Respect default rules
-
-    reports {
-        html {
-            required.set(true)
-            outputLocation.set(layout.buildDirectory.file("reports/detekt/detekt-report.html"))
-        }
-
-        xml.required.set(false)
-        txt.required.set(false)
-        sarif.required.set(false)
-        md.required.set(false)
-    }
-
-    finalizedBy(detektMessage)
-}
-
-val detektMessage by tasks.register("HOW TO FIX DETEKT FAIL") {
-    doLast {
-        val reportPath = "${layout.buildDirectory.get()}/reports/detekt/detekt-report.html"
-        val reportUrl = "file://${File(reportPath).absolutePath}"
-
-        println("\nDetekt task ended")
-        println("Report address (copy and paste to web-browser):")
-        println(reportUrl)
-
-        if (tasks.named("detekt").get().state.failure != null) {
-            autoOpenHtmlReport(reportPath)
-            throw GradleException("""
-                
-            ________________________________________________________________________________________   
-            📃FULL REPORT (🌐 copy and paste to web-browser):
-            $reportPath
-            ________________________________________________________________________________________
-            🔧 HOW TO FIX? 
-            1. Open the report and fix the listed issues.
-            2. Put changes.
-            3. Re-build project (preferably) OR run locally (terminal/Git Bash)
-                    ./gradlew detekt
-            
-            Optionally (pre-debug mode): Try locally (terminal/Git Bash)
-                    ./gradlew build -x detekt -x lint
-            to see app running without detekt and lint
-
-            ⚠️ The default build will fail until all issues are resolved.
-            ________________________________________________________________________________________
-            
-            """.trimIndent())
-        }
-    }
-}
-
-afterEvaluate {
-    tasks.named("assembleDebug") {
-        finalizedBy("test")
-        finalizedBy("detekt")
-        finalizedBy("lint")
-    }
-}
-
-tasks.withType<Test> {
-    finalizedBy("openTestReportOnFailure")
-}
-
-tasks.register("openTestReportOnFailure") {
-    onlyIf {
-        tasks.withType<Test>().any { it.state.failure != null }
-    }
-    doLast {
-        val reportPath = "${projectDir}/build/reports/tests/testDebugUnitTest/index.html"
-        autoOpenHtmlReport(reportPath)
-    }
-}
-
 dependencies {
     implementation(project(":core:ui"))
 
@@ -162,25 +66,6 @@ dependencies {
     testImplementation(platform(libs.junit.bom))
     testRuntimeOnly(libs.junit.platform.launcher)
     testImplementation(libs.mockk)
-
-    detektPlugins(libs.detekt.formatting)
-    testImplementation(libs.detekt.test)
-}
-
-fun autoOpenHtmlReport (reportPath: String) {
-    val reportUrl = "file://${file(reportPath).absolutePath}"
-    try {
-        val os = System.getProperty("os.name").lowercase()
-
-        when {
-            os.contains("win") -> ProcessBuilder("rundll32", "url.dll", "FileProtocolHandler", reportUrl).start()
-            os.contains("mac") -> ProcessBuilder("open", reportUrl).start()
-            os.contains("nix") || os.contains("nux") -> ProcessBuilder("xdg-open", reportUrl).start()
-            else -> println("Unsupported OS for automatic opening")
-        }
-    } catch (e: Exception) {
-        println("Could not open browser automatically: ${e.message}")
-    }
 }
 
 tasks.withType<Test> {
