--- app/detekt-setting/detekt.yml
@@ -1,24 +1,0 @@
-style:
-  MagicNumber:
-    active: true
-    excludes:
-      - '**/Color.kt'
-      - '**/test/**'
-      - '**/ExampleUnitTest.kt'
-      - '**/example/**'
-  NewLineAtEndOfFile:
-    active: false
-
-formatting:
-  FinalNewline:
-    active: false
-  NewLineAtEndOfFile:
-    active: false
-
-naming:
-  FunctionNaming:
-    active: false
-
-empty-blocks:
-  EmptyDefaultConstructor:
-    active: false
