+++ gradle.properties
@@ -22,3 +22,6 @@ kotlin.code.style=official
 # thereby reducing the size of the R class for that library
 android.nonTransitiveRClass=true
 android.defaults.buildfeatures.buildconfig=true
+
+# Disable Lint
+gradle=build -x lint -x lintVitalRelease
